Rename MinPlasCalc to minplascalc
