+++ minplascalc.py
